feat(platforms): start moving oauth loginc to providers
