+++ platforms/src/Github/index.ts
@@ -1,5 +1,6 @@
 export { GithubPlatform } from "./App-Bindings";
 export { GithubProviderConfig } from "./Providers-config";
+export { GithubPlatformDetails } from "./Providers-config";
 export {
   TenOrMoreGithubFollowers,
   FiftyOrMoreGithubFollowers,
