+++ platforms/src/Google/app.js
@@ -1,0 +1,2 @@
+export { GooglePlatform } from "./App-Bindings";
+export { GoogleProviderConfig } from "./Providers-config";
