+++ platforms/src/Google/app/index.js
@@ -1,0 +1,3 @@
+export { GooglePlatform } from "./App-Bindings";
+export { GoogleProviderConfig } from "./Providers-config";
+// export { GoogleProvider } from "./Providers/google";
