--- platforms/src/Zksync/index.ts
@@ -1,4 +1,0 @@
-export { ZkSyncProvider } from "./Providers/zkSync";
-
-export { ZkSyncPlatformDetails, ZkSyncProviderConfig } from "./Providers-config";
-export { ZkSyncPlatform } from "./App-Bindings";
