+++ platforms/src/app/index.js
@@ -1,0 +1,9 @@
+// export * as Twitter from "../Twitter/app";
+// export * as Ens from "../Ens/app";
+// export * as Github from "../Github/app";
+// export * as Gitcoin from "../Gitcoin/app";
+// export * as Lens from "../Lens/app";
+// export * as Poh from "../Poh/app";
+// export * as Google from "../Google/app";
+export { GooglePlatform } from "../Google/App-Bindings";
+export { GoogleProviderConfig } from "../Google/Providers-config";
